+++ lib/phonelib/phone.rb
@@ -1,26 +1,25 @@
 module Phonelib
   # class for parsed phone number, includes validation and formatting methods
   class Phone
-    # defining reader methods for class variables
-    # original phone number passed for parsing
+    # @!attribute [r] original
+    # @return [String] original phone number passed for parsing
     attr_reader :original
-    # phone extension passed for parsing after a number
+    # @!attribute [r] extension
+    # @return [String] phone extension passed for parsing after a number
     attr_reader :extension
 
     # including module that has all phone analyzing methods
     include Phonelib::PhoneAnalyzer
     include Phonelib::PhoneExtendedData
+    include Phonelib::PhoneFormatter
 
     # class initialization method
-    #
-    # ==== Attributes
-    #
-    # * +phone+   - Phone number for parsing
-    # * +country+ - Country specification for parsing. Must be ISO code of
+    # @param phone [String] Phone number for parsing
+    # @param country [String|Symbol] Country specification for parsing. Must be ISO code of
     #   country (2 letters) like 'US', 'us' or :us for United States
-    #
-    def initialize(original, country = nil)
-      @original, @extension = separate_extension(original.to_s)
+    # @return [Phonelib::Phone] parsed phone instance
+    def initialize(phone, country = nil)
+      @original, @extension = separate_extension(phone.to_s)
       @extension.gsub!(/[^0-9]/, '') if @extension
 
       if sanitized.empty?
@@ -33,6 +32,7 @@ module Phonelib
     end
 
     # method to get sanitized phone number (only numbers)
+    # @return [String] Sanitized phone number
     def sanitized
       @sanitized = if Phonelib.strict_check
                      @original
@@ -42,36 +42,43 @@ module Phonelib
     end
 
     # Returns all phone types that matched valid patterns
+    # @return [Array] all valid phone types
     def types
       @data.flat_map { |iso2, data| data[:valid] }.uniq
     end
 
     # Returns all possible types that matched possible patterns
+    # @return [Array] all possible phone types
     def possible_types
       @data.flat_map { |iso2, data| data[:possible] }.uniq
     end
 
     # Returns first phone type that matched
+    # @return [Symbol] valid phone type
     def type
       types.first
     end
 
     # Returns human representation of all matched phone types
+    # @return [Array] Array of human readable valid phone types
     def human_types
       types.map { |type| Core::TYPES_DESC[type] }
     end
 
     # Return human representation of phone type
+    # @return [String] Human readable valid phone type
     def human_type
       Core::TYPES_DESC[type]
     end
 
     # Returns all countries that matched valid patterns
+    # @return [Array] Possible ISO2 country codes array
     def countries
       @data.map { |iso2, data| iso2 }
     end
 
     # Return countries with valid patterns
+    # @return [Array] Valid ISO2 country codes array
     def valid_countries
       @valid_countries ||= countries.select do |iso2|
         @data[iso2][:valid].any?
@@ -79,16 +86,19 @@ module Phonelib
     end
 
     # Return valid country
+    # @return [String] valid ISO2 country code
     def valid_country
-      @valid_country ||= get_main_country(valid_countries)
+      @valid_country ||= main_country(valid_countries)
     end
 
     # Returns first country that matched valid patterns
+    # @return [String] valid country ISO2 code or first matched country ISO2 code
     def country
-      @country ||= valid_country || get_main_country(countries)
+      @country ||= valid_country || main_country(countries)
     end
 
     # Returns the country code from the original phone number.
+    # @return [String] matched country phone code
     def country_code
       if country_data = Phonelib.phone_data[country]
         country_data[:country_code]
@@ -96,36 +106,52 @@ module Phonelib
     end
 
     # Returns whether a current parsed phone number is valid
+    # @return [Boolean] parsed phone is valid
     def valid?
       @data.select { |iso2, data| data[:valid].any? }.any?
     end
 
     # Returns whether a current parsed phone number is invalid
+    # @return [Boolean] parsed phone is invalid
     def invalid?
       !valid?
     end
 
     # Returns whether a current parsed phone number is possible
+    # @return [Boolean] parsed phone is possible
     def possible?
       @data.select { |iso2, data| data[:possible].any? }.any?
     end
 
     # Returns whether a current parsed phone number is impossible
+    # @return [Boolean] parsed phone is impossible
     def impossible?
       !possible?
     end
 
     # returns area code of parsed number
+    # @return [String|nil] parsed phone area code if available
     def area_code
       return nil unless possible?
-      format_match, format_string = get_formatting_data
 
-      if format_string =~ /^.*[0-9]+.*\$1/ && format_match
-        format_string.gsub(/\$1.*$/, format_match[1]).gsub(/[^\d]+/, '')
+      # has national prefix
+      return nil unless @data[country][Core::NATIONAL_PREFIX] || country == 'IT'
+      # fixed or mobile
+      return nil unless Core::AREA_CODE_TYPES.include?(type)
+      # mobile && mexico, argentina, brazil
+      return nil if type == Core::MOBILE && !Core::AREA_CODE_MOBILE_COUNTRIES.include?(country)
+
+      format_match, _format_string = get_formatting_data
+      take_group = 1
+      if type == Core::MOBILE && Core::AREA_CODE_MOBILE_TOKENS[country]
+        format_match[1] == Core::AREA_CODE_MOBILE_TOKENS[country]
+        take_group = 2
       end
+      format_match[take_group]
     end
 
     # returns local number
+    # @return [String] local number
     def local_number
       return national unless possible?
       format_match, format_string = get_formatting_data
@@ -138,65 +164,11 @@ module Phonelib
       end
     end
 
-    # Returns formatted national number
-    def national(formatted = true)
-      return @national_number unless valid?
-      format_match, format_string = get_formatting_data
-
-      if format_match
-        out = format_string.gsub(/\$\d/) { |el| format_match[el[1].to_i] }
-        formatted ? out : out.gsub(/[^0-9]/, '')
-      else
-        @national_number
-      end
-    end
-
-    # Returns e164 formatted phone number
-    def international(formatted = true)
-      return nil if sanitized.nil? || sanitized.empty?
-      return "+#{country_prefix_or_not}#{sanitized}" unless valid?
-      return "#{@data[country][Core::COUNTRY_CODE]}#{@national_number}" unless formatted
-
-      format = @data[country][:format]
-      if matches = @national_number.match(/#{format[Core::PATTERN]}/)
-        fmt = format[:intl_format] || format[:format]
-        national = fmt.gsub(/\$\d/) { |el| matches[el[1].to_i] }
-      else
-        national = @national_number
-      end
-
-      "+#{country_code} #{national}"
-    end
-
-    # returns national formatted number with extension added
-    def full_national
-      "#{national}#{formatted_extension}"
-    end
-
-    # returns international formatted number with extension added
-    def full_international
-      "#{international}#{formatted_extension}"
-    end
-
-    # returns e164 format of phone with extension added
-    def full_e164
-      "#{e164}#{formatted_extension}"
-    end
-
-    # Returns e164 unformatted phone number
-    def e164
-      international = self.international
-      international and international.gsub /[^+0-9]/, ''
-    end
-
     # Returns whether a current parsed phone number is valid for specified
     # country
-    #
-    # ==== Attributes
-    #
-    # * +country+ - ISO code of country (2 letters) like 'US', 'us' or :us
+    # @param country [String|Symbol] ISO code of country (2 letters) like 'US', 'us' or :us
     #   for United States
-    #
+    # @return [Boolean] parsed phone number is valid
     def valid_for_country?(country)
       country = country.to_s.upcase
       @data.find do |iso2, data|
@@ -206,62 +178,27 @@ module Phonelib
 
     # Returns whether a current parsed phone number is invalid for specified
     # country
-    #
-    # ==== Attributes
-    #
-    # * +country+ - ISO code of country (2 letters) like 'US', 'us' or :us
+    # @param country [String|Symbol] ISO code of country (2 letters) like 'US', 'us' or :us
     #   for United States
-    #
+    # @return [Boolean] parsed phone number is invalid
     def invalid_for_country?(country)
       !valid_for_country?(country)
     end
 
     private
 
-    # defines whether to put country prefix or not
-    def country_prefix_or_not
-      return '' unless country_code
-      sanitized.start_with?(country_code) ? '' : country_code
-    end
-
-    # returns extension with separator defined
-    def formatted_extension
-      return '' if @extension.nil? || @extension.empty?
-
-      "#{Phonelib.extension_separator}#{@extension}"
-    end
-
-    # extracts extension from passed phone number if provided
+    # @private extracts extension from passed phone number if provided
     def separate_extension(original)
       regex = cr("[#{Phonelib.extension_separate_symbols}]")
       split = (original || '').split regex
       [split.first || '', split[1..-1] && split[1..-1].join || '']
     end
 
-    # get main country for code among provided countries
-    def get_main_country(countries_array)
+    # @private get main country for code among provided countries
+    def main_country(countries_array)
       countries_array.find do |iso2|
         @data[iso2][Core::MAIN_COUNTRY_FOR_CODE] == 'true'
       end || countries_array.first
-    end
-
-    # Get needable data for formatting phone as national number
-    def get_formatting_data
-      return @formatting_data if @formatting_data
-
-      format = @data[country][:format]
-      prefix = @data[country][Core::NATIONAL_PREFIX]
-      rule = (format[Core::NATIONAL_PREFIX_RULE] ||
-          @data[country][Core::NATIONAL_PREFIX_RULE] || '$1')
-
-      # change rule's constants to values
-      rule.gsub!(/(\$NP|\$FG)/, '$NP' => prefix, '$FG' => '$1')
-
-      # add space to format groups, change first group to rule,
-      format_string = format[:format].gsub(/(\d)\$/, '\\1 $').gsub('$1', rule)
-
-      @formatting_data =
-          [@national_number.match(/#{format[Core::PATTERN]}/), format_string]
     end
   end
 end