+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Upcoming Changes (unreleased)
+- fixing to override default country when + passed in the beginning of phone
 
 ## 0.6.2 - June 14, 2016
 - fixed bug in ```international``` method when no country can be defined