+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Upcoming Changes (unreleased)
+
+## 0.6.3 - August 8, 2016
 - fixing to override default country when + passed in the beginning of phone
 - fixed error when int passed as phone number
 - fixed error when ```";"``` passed as phone number